% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlleleCounts.R
\name{allele_loader}
\alias{allele_loader}
\title{\code{allele_loader} internal wrapper function for choosing the right read-in function}
\usage{
allele_loader(path, type, compressed)
}
\arguments{
\item{path}{character string input containing the path to the directory containing the
input files; helper function used in readAlleleCounts()}

\item{type}{character vector containing options for the input data_type}

\item{compressed}{binary classification if the input data are .gz compressed}
}
\value{
depending on the chosen type, fitting read_in function will be used
}
\description{
\code{allele_loader} internal wrapper function for choosing the right read-in function
}
