% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlleleCounts.R
\name{check_for_compressed}
\alias{check_for_compressed}
\title{\code{.check_for_compressed()} internal function that checks whether the input files are gzipped}
\usage{
check_for_compressed(path, compressed, error = TRUE)
}
\arguments{
\item{path}{file path of the directory containing the input files as character string; taken from
readAllleleCounts()}

\item{compressed}{binary variable taken from readAlleleCounts() indicating if the user files are gzipped}

\item{error}{binary variable taken from readAlleleCounts() indicating if there should be an error message, when the gzipped files
cannot be found in the given directory}
}
\value{
character string adding ".gz" to the filenames
}
\description{
\code{.check_for_compressed()} internal function that checks whether the input files are gzipped
}
