% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlleleCounts.R
\name{check_txt_tsv}
\alias{check_txt_tsv}
\title{\code{check_txt_tsv} internal function that checks if the input files are .txt or .tsv files; might be needed when
spatial data get incorporated}
\usage{
check_txt_tsv(path, filename)
}
\arguments{
\item{path}{file path of the directory containing the input files as character string; taken from
readAllleleCounts()}

\item{filename}{character string containing the names of the files to be checked}
}
\value{
file name either appended with .txt or .tsv - depending on what could be found
in the input directory
}
\description{
\code{check_txt_tsv} internal function that checks if the input files are .txt or .tsv files; might be needed when
spatial data get incorporated
}
