% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAlleleExperiment.R
\name{check_unknowns}
\alias{check_unknowns}
\title{Internal Error handler}
\usage{
check_unknowns(sce, find_allele_ids)
}
\arguments{
\item{sce}{A \code{\link{SingleCellExperiment}} object.}

\item{find_allele_ids}{A list containing allele identifiers present in the raw data.}
}
\value{
Error message if condition is not met.
}
\description{
Internal function used in \code{get_allelecounts()} to check if an allele identifier cannot be found in the lookup table AND does not have proper nomenclature
form, then the execution of further step and thus the generation of an SingleCellAlleleExperiment object stopped.
}
