% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAlleleExperiment.R
\name{filter_norm}
\alias{filter_norm}
\title{Preprocessing}
\usage{
filter_norm(sce, threshold = 0)
}
\arguments{
\item{sce}{A \code{\link{SingleCellExperiment}} object.}

\item{threshold}{An integer value used as a threshold for filtering low-quality barcodes/cells.}
}
\value{
A SingleCellExperiment object.
}
\description{
Internal function used in \code{SingleCellAlleleExperiment()} constructor as a preprocessing step for
filtering the barcodes and normalizing the count values.
}
