% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{find_allele_ids_amp}
\alias{find_allele_ids_amp}
\title{Identify rows containing allele information for Amplicon}
\usage{
find_allele_ids_amp(sce)
}
\arguments{
\item{sce}{singlecellexperiment object}
}
\value{
list of rownames starting with "HLA"
}
\description{
Identify rows containing allele information for Amplicon
}
