% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{find_allele_ids_ens}
\alias{find_allele_ids_ens}
\title{Identify rows containing allele information for WTA}
\usage{
find_allele_ids_ens(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{
subsample of the scae containing all rows with allele_counts
}
\description{
Internal function used in \code{get_allelecounts()} to subsample the quantification assay and only
return the rows specifying allele-quantification information.
}
