% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAlleleExperiment.R
\name{find_not_ident}
\alias{find_not_ident}
\title{Find not yet known allele identifiers}
\usage{
find_not_ident(sce, agene_names)
}
\arguments{
\item{sce}{A \code{\link{SingleCellExperiment}} object.}

\item{agene_names}{A list of allele gene names.}
}
\value{
A list of character strings of identifiers that can not be found in the allele lookup table.
}
\description{
Internal function used in \code{get_allelecounts()}to find allele identifier that are not present in the lookup table.
}
