% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{get_agenes}
\alias{get_agenes}
\title{\code{get_agenes} getter function which creates scae-object for the subassay containing allele_gene information}
\usage{
get_agenes(scae)
}
\arguments{
\item{scae}{SingleCellAlleleExperiment object}
}
\value{
subsampled SingleCellAlleleExperiment object
}
\description{
\code{get_agenes} getter function which creates scae-object for the subassay containing allele_gene information
}
