% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAlleleExperiment.R
\name{get_allelecounts}
\alias{get_allelecounts}
\title{Get Subassay with allele gene names and raw allele quantification}
\usage{
get_allelecounts(sce, lookup)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object.}

\item{lookup}{A data.frame object containing the lookup table.}
}
\value{
A SingleCellExperiment object
}
\description{
Internal function used to build a subassay containing counts from raw alleles.
The rownames  of this subassay are already translated to the corresponding
immune gene identifier, which are extracted from the lookup table.
}
