% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{get_alleles}
\alias{get_alleles}
\title{\code{get_alleles} getter function which creates scae-object for the subassay containing allele information}
\usage{
get_alleles(scae)
}
\arguments{
\item{scae}{SingleCellAlleleExperiment object}
}
\value{
subsampled SingleCellAlleleExperiment object
}
\description{
\code{get_alleles} getter function which creates scae-object for the subassay containing allele information
}
