% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlleleCounts.R
\name{get_knee_info}
\alias{get_knee_info}
\title{Knee plot info}
\usage{
get_knee_info(matrix, genes, barcodes)
}
\arguments{
\item{matrix}{A sparse \code{\link[Matrix]{Matrix}} object containing
the quantification data.}

\item{genes}{A data.frame object containing gene identifiers.}

\item{barcodes}{A data.frame object containing barcode identifiers.}
}
\value{
A list including a data.frame with barcode rank information,
the corresponding knee and inflection point.
}
\description{
Creates a knee plot information, ranking the barcodes according to their
total UMI count. The information is later on passed to the
\code{metadata(scae)[["knee_info"]]} slot.
}
