% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_unknown}
\alias{get_unknown}
\title{Get unknown allele rows}
\usage{
get_unknown(scae)
}
\arguments{
\item{scae}{SingleCellAlleleExperiment object}
}
\value{
subsampled SingleCellAlleleExperiment object
}
\description{
Getter function returning subsampled SCAE object with all rows containing unknown allele information thats present in the data and in the proper nomenclature.
These rows are identified by "I" in rowData(scae)$NI_I and "A_unknown" in rowData(scae)$Quant_type.
}
