% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{log_transform}
\alias{log_transform}
\title{log transform normalized counts}
\usage{
log_transform(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{
SingleCellAlleleExperiment object with an additional assay containing
log normalized counts
}
\description{
Internal function used in in
}
