% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{lookup_gene}
\alias{lookup_gene}
\title{\code{lookup_gene} internal function adding information to the SingleCellAlleleExperiment object by
filling the rowData columns ("NI_I" and "Quant_type") with the correct information for the rows of the newly added
subassays after the}
\usage{
lookup_gene(sce, lookup)
}
\arguments{
\item{sce}{singlecellexperiment object}

\item{lookup}{allele lookup table}
}
\value{
singlecellexperiment object
}
\description{
\code{lookup_gene} internal function adding information to the SingleCellAlleleExperiment object by
filling the rowData columns ("NI_I" and "Quant_type") with the correct information for the rows of the newly added
subassays after the
}
