% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlleleCounts.R
\name{read_from_sparse_allele}
\alias{read_from_sparse_allele}
\title{Reading in allele-aware quantification data}
\usage{
read_from_sparse_allele(
  path,
  exp_type = exp_type,
  barcode_file,
  gene_file,
  matrix_file
)
}
\arguments{
\item{path}{character string input containing the path to the directory containing the
input files}

\item{exp_type}{either \code{WTA} or \code{Amplicon} depending on the used experiments technology}

\item{barcode_file}{character string determining the name of the file containing the barcode identifiers}

\item{gene_file}{character string determining the name of the file containing the feature identifiers}

\item{matrix_file}{character string determining the name of the file containing the count matrix}
}
\value{
list with the read_in data sorted into different slots
}
\description{
Internal function used in \code{readAlleleCounts()} that reads in the data stated in the given directory path.
}
