% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlleleCounts.R
\name{read_Lookup}
\alias{read_Lookup}
\title{Read in allele lookup}
\usage{
read_Lookup(path, exp_type, lookup_file)
}
\arguments{
\item{path}{A character string determining the path to the directory containing the input files.}

\item{exp_type}{A vector containing two character strings. Either \code{"WTA"} or \code{"Amplicon"} are valid inputs. Choose one depending on the used transcriptomics approach.}

\item{lookup_file}{A character string determining the name of the lookup table file.}
}
\value{
A data.frame containing a representation of the lookup table.
}
\description{
Internal function used in \code{readAlleleCounts()} to read in the allele lookup table.
}
