% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlleleCounts.R
\name{readAlleleCounts}
\alias{readAlleleCounts}
\title{Reading in allele quantification data into SingleCellAlleleExperiment object}
\usage{
readAlleleCounts(
  samples,
  sample.names = names(samples),
  filter_threshold = 0,
  exp_type = c("WTA", "Amplicon"),
  symbols = NULL,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{samples}{character string input containing the path to the directory containing the
input files}

\item{sample.names}{character string for a sample_name identifier}

\item{filter_threshold}{count threshold for filtering barcodes/cells}

\item{exp_type}{either \code{WTA} or \code{Amplicon} depending on the used experiments technology}

\item{symbols}{identifier used to choose which database-function to use to retrieve the ncbi gene names}

\item{BPPARAM}{A BiocParallelParam object specifying how loading should be parallelized for multiple samples}
}
\value{
SingleCellAlleleExperiment object
}
\description{
Main read in function for reading in given allele quantification data and
loading the data into an \code{SingleCellAlleleExperiment} object. Input data are stored in a shared folder.
Expected naming scheme of the files:
\itemize{
\item quantification matrix: \code{matrix.mtx}
\item barcode information: \code{barcodes.txt}
\item feature information: \code{features.txt}
\item allele lookup table: \code{lookup_table_HLA_only}
}
}
