% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlleleCounts.R
\name{readLookup}
\alias{readLookup}
\title{Read in lookup table}
\usage{
readLookup(path, exp_type)
}
\arguments{
\item{path}{character string of the file path to the directory containing the expected input files}

\item{exp_type}{character string to pass information about the used experimental approach. Either use \code{exp_type = "WTA"} or \code{exp_type = "Amplicon"}.}
}
\value{
lookup table
}
\description{
Internal function used in \code{readAlleleCounts()} to read in the lookup table.
}
