% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{rowData<-,SingleCellAlleleExperiment-method}
\alias{rowData<-,SingleCellAlleleExperiment-method}
\title{rowData setter}
\usage{
\S4method{rowData}{SingleCellAlleleExperiment}(x, ...) <- value
}
\arguments{
\item{x}{A SingleCellAlleleExperiment object}

\item{...}{additional paramaters passed to rowData<-}

\item{value}{arbitrary value (can be integer, character,..)}
}
\value{
SingleCellAlleleExperiment object
}
\description{
rowData setter
}
