% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAlleleExperiment-rowData.R
\name{rowData<-,SingleCellAlleleExperiment,NULL-method}
\alias{rowData<-,SingleCellAlleleExperiment,NULL-method}
\title{rowData-NULL-setter for the SingleCellAlleleExperiment class}
\usage{
\S4method{rowData}{SingleCellAlleleExperiment,NULL}(x) <- value
}
\arguments{
\item{x}{A \code{\link{SingleCellAlleleExperiment}} object}

\item{value}{NULL}
}
\value{
A \code{\link{SingleCellAlleleExperiment}} object
}
\description{
Setter function for the rowData slot for the \code{\link{SingleCellAlleleExperiment}} class.
}
