% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{scae_subset_alleles}
\alias{scae_subset_alleles}
\title{Get allele rows}
\usage{
scae_subset_alleles(scae)
}
\arguments{
\item{scae}{A \code{\link{SingleCellAlleleExperiment}} object.}
}
\value{
A SingleCellAlleleExperiment object.
}
\description{
Getter function returning subsampled SCAE object with all rows containing raw allele information. These rows
are identified by "I" in rowData(scae)$NI_I and "A" in rowData(scae)$Quant_type.
}
