% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{scae_subset_functional}
\alias{scae_subset_functional}
\title{Get functional class rows}
\usage{
scae_subset_functional(scae)
}
\arguments{
\item{scae}{A \code{\link{SingleCellAlleleExperiment}} object.}
}
\value{
A SingleCellAlleleExperiment object.
}
\description{
Getter function returning subsampled SCAE object with all rows containing functional class information. These rows are
identified by "I" in rowData(scae)$NI_I and "F" in rowData(scae)$Quant_type.
}
\examples{

example_data_5k <- scaeData::scaeDataGet(dataset = "pbmc_5k")

scae <- read_allele_counts(example_data_5k$dir,
                          sample_names = "example_data_wta",
                          filter = "custom",
                          exp_type = "WTA",
                          lookup_file = "lookup_table_HLA_only.csv",
                          barcode_file = example_data_5k$barcodes,
                          gene_file = example_data_5k$features,
                          matrix_file = example_data_5k$matrix,
                          filter_threshold = 0,
                          example_dataset = TRUE,
                          verbose = TRUE)

scae

rownames(scae_subset_functional(scae))

scae_functional_class <- scae_subset_functional(scae)

scae_functional_class

rownames(scae_functional_class)

}
