% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{sfactors}
\alias{sfactors}
\title{Computing size factors and normalize the filtered counts}
\usage{
sfactors(sce, num_of_bins = 75)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{num_of_bins}{number of bins for the histograms of sizeFactors; default set to 75}
}
\value{
normalized SingleCellExperiment object
}
\description{
Internal function used in \code{filter_norm()}
}
