% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\docType{class}
\name{SingleCellAlleleExperiment-class}
\alias{SingleCellAlleleExperiment-class}
\alias{.scae}
\title{\code{.scae} SingleCellAlleleExperiment-Class derived from SingleCellExperiment-class}
\value{
definition for the scae class
}
\description{
\code{.scae} SingleCellAlleleExperiment-Class derived from SingleCellExperiment-class
}
