% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAlleleExperiment-methods.R
\name{SingleCellAlleleExperiment-misc}
\alias{SingleCellAlleleExperiment-misc}
\alias{show,SingleCellAlleleExperiment-method}
\title{Miscellaneous SingleCellAlleleExperiment methods}
\usage{
\S4method{show}{SingleCellAlleleExperiment}(object)
}
\arguments{
\item{object}{a \code{\link{SingleCellAlleleExperiment}} object}
}
\value{
Returns NULL
}
\description{
Miscellaneous methods for the \code{\link{SingleCellAlleleExperiment}} class
and its descendants that do not fit into any other documentation category
such as, for example, show methods.
}
