% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{add_sample_tags}
\alias{add_sample_tags}
\title{adding sample tag information to colData}
\usage{
add_sample_tags(path, scae)
}
\arguments{
\item{path}{character string input containing the path to the directory containing the
input files}

\item{scae}{SingleCellAlleleExperiment object}
}
\value{
updated SingleCellAlleleExperiment object
}
\description{
adding sample tag information to colData
}
