% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAlleleExperiment.R
\name{add_sample_tags}
\alias{add_sample_tags}
\title{Adding sample tag information to colData}
\usage{
add_sample_tags(path, scae, tag_feature_mtx, tag_feature_barcodes)
}
\arguments{
\item{path}{character string input containing the path to the directory containing the
input files}

\item{scae}{A \code{\link{SingleCellAlleleExperiment}} object.}

\item{tag_feature_mtx}{A character string determining the name of the file containing the sample-tag quantification data.}

\item{tag_feature_barcodes}{A character string determining the name of the file containing the sample-tag barcode identifiers.}
}
\value{
A SingleCellAlleleExperiment object.
}
\description{
Internal function used in 'readAlleleCounts()'. Stated here because its supposed to be a transformation step of the SCAE object.
Adding sample tag information to colData
}
