% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{add_sample_tags}
\alias{add_sample_tags}
\title{Adding sample tag information to colData}
\usage{
add_sample_tags(path, scae, tag_feature_mtx, tag_feature_barcodes)
}
\arguments{
\item{path}{character string input containing the path to the directory containing the
input files}

\item{scae}{SingleCellAlleleExperiment object}

\item{tag_feature_mtx}{add text here}

\item{tag_feature_barcodes}{add text here}
}
\value{
updated SingleCellAlleleExperiment object
}
\description{
Internal function used in 'readAlleleCounts()'. Stated here because its supposed to be a transformation step of the SCAE object.
Adding sample tag information to colData
}
