% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{alleles2genes}
\alias{alleles2genes}
\title{\code{allele2genes} first internal translation function used in the SingleCellAlleleExperiment constructor
computing the one of the two new subassay that get appended to the
quantification assay. This subassay contains the allele-gene names instead of the allele-identifiers and
sums up the expression counts of alleles that have the same allele-gene name.}
\usage{
alleles2genes(sce, lookup)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{lookup}{allele lookup table}
}
\value{
adds allele-gene subassay containing summarized count-information for the allele-genes to the main assay
}
\description{
\code{allele2genes} first internal translation function used in the SingleCellAlleleExperiment constructor
computing the one of the two new subassay that get appended to the
quantification assay. This subassay contains the allele-gene names instead of the allele-identifiers and
sums up the expression counts of alleles that have the same allele-gene name.
}
