% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{alleles2genes}
\alias{alleles2genes}
\title{Building first new subassay for SingleCellAllelexperiment object}
\usage{
alleles2genes(sce, lookup, exp_type)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{lookup}{lookup table info passed from \code{SingleCellAlleleExperiment()}.}

\item{exp_type}{character string to pass information about the used experimental approach. Either use \code{exp_type = "WTA"} or \code{exp_type = "Amplicon"}, passed from the \code{readAlleleCounts()} function.}
}
\value{
adds allele gene-subassay containing summarized count information for the allele genes
}
\description{
Internal function for the first assay extension used in the \code{SingleCellAlleleExperiment()} constructor
computing the first of the two new subassays that get appended to the
quantification assay. This subassay contains the allele gene identifiers instead of the allelen identifiers and
sums up the expression counts of alleles that have the same allele gene identifiers.
}
