% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{alleles2genes}
\alias{alleles2genes}
\title{Building first new subassay for SingleCellAllelexperiment object}
\usage{
alleles2genes(sce, lookup)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{lookup}{allele lookup table}
}
\value{
adds allele gene-subassay containing summarized count information for the allele genes
}
\description{
Internal function for the first assay extension used in the \code{SingleCellAlleleExperiment()} constructor
computing the first of the two new subassays that get appended to the
quantification assay. This subassay contains the allele gene identifiers instead of the allelen identifiers and
sums up the expression counts of alleles that have the same allele gene identifiers.
}
