% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{check_unknowns}
\alias{check_unknowns}
\title{Internal error handler for unknown allele identifiers}
\usage{
check_unknowns(sce, find_allele_ids)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{find_allele_ids}{list containing character strings of allele identifiers present in the raw data}
}
\value{
a message and stops code execution if condition not met
}
\description{
Internal function used in \code{get_allelecounts()} to check if an allele identifier can not be found in the lookup table AND does not have proper nomenclature
form, then the execution of further step and thus the generation of an SingleCellAlleleExperiment object stops, giving info about which allele identifiers are invalid.
}
