% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAlleleExperiment.R
\name{cutname}
\alias{cutname}
\title{Build new substring}
\usage{
cutname(allele_id)
}
\arguments{
\item{allele_id}{A list of character strings for unidentified allele_names that have proper \href{https://hla.alleles.org/nomenclature/index.html}{allele-nomenclature}.}
}
\value{
A list of charcter strings resembling HLA gene names.
}
\description{
Internal function used in \code{get_allelecounts()}. Function is used to cut character string at "*" character and return it. Only used to cut the names
if unknown alleles, as the lookup table does not provide any corresponding immune gene name.
}
