% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{cutname}
\alias{cutname}
\title{Build new substring}
\usage{
cutname(allele_id)
}
\arguments{
\item{allele_id}{list of unidentified allele_names that are still in a proper nomenclature form}
}
\value{
new identifier with a cut off name to be present in the list of gene_names
}
\description{
Internal function used in \code{get_allelecounts()}. Function is used to cut character string at "*" character and return it.
}
