% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{ext_rd}
\alias{ext_rd}
\title{Extending rowData}
\usage{
ext_rd(sce, exp_type, symbols)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{exp_type}{either \code{"WTA"} or \code{"Amplicon"} depending on the used experiments technology}

\item{symbols}{identifier used to choose which database-function to use to retrieve the NCBI gene names}
}
\value{
SingleCellAlleleExperiment object with extended rowData
}
\description{
Internal function used in the \code{SingleCellAlleleExperiment()} constructor adding information to the SingleCellAlleleExperiment object by
extending the rowData by two columns. \code{NI_I} is a classifier for each feature_row if its considered a
non-immune (NI) or immune (I) gene. \code{Quant_type} is a classifier for determining which row is related to which
subassay of the extended main assay in the \code{SingleCellAlleleExperiment}. "A" corresponds to allele, "G" to allele gene and
"F" to functional allele class.
}
