% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAlleleExperiment.R
\name{ext_rd}
\alias{ext_rd}
\title{Extending rowData}
\usage{
ext_rd(sce, exp_type, verbose = FALSE)
}
\arguments{
\item{sce}{A \code{\link{SingleCellExperiment}} object. Object is initally constructed in the \code{SingleCellAlleleExperiment} constructor.}

\item{exp_type}{A vector containing two character strings. Either \code{"WTA"} or \code{"Amplicon"} are valid inputs. Choose one depending on the used transcriptomics approach.}

\item{verbose}{A logical parameter to decide if runtime-messages should be shown during function execution.
Use \code{FALSE} if no info runtime-messages should be shown (default), and \code{TRUE} for showing runtime-messages.}
}
\value{
A SingleCellExperiment object.
}
\description{
Internal function used in the \code{SingleCellAlleleExperiment()} constructor adding information to the SingleCellAlleleExperiment object by
extending the rowData by two columns. \code{NI_I} is a classifier for each feature_row if its considered a
non-immune (NI) or immune (I) gene. \code{Quant_type} is a classifier for determining which row is related to which
subassay of the extended main assay in the \code{SingleCellAlleleExperiment}. "A" corresponds to allele, "G" to allele gene and
"F" to functional allele class.
}
