% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{filter_norm}
\alias{filter_norm}
\title{Preprocessing}
\usage{
filter_norm(sce, threshold = 0)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{threshold}{positive integer value used as count threshold for filtering barcodes/cells. Default value is set at \code{filter_threshold = 0} in the \code{readAlleleCounts()} function. Value passed from \code{SingleCellAlleleExperiment()}.}
}
\value{
filtered and normalized SingleCellExperiment object
}
\description{
Internal function used in \code{SingleCellAlleleExperiment()} constructor as a preprocessing step for
filtering the barcodes and normalizing the count values. Computes the scaling factors for each cell later used for normalization.
}
