% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{filter_norm}
\alias{filter_norm}
\title{Preprocessing}
\usage{
filter_norm(sce, threshold = 0)

filter_norm(sce, threshold = 0)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{threshold}{Counts threshold for barcodes}
}
\value{
filtered and normalized SingleCellExperiment object

filtered and normalized SingleCellExperiment object
}
\description{
Internal function used in \code{SingleCellAlleleExperiment()} constructor as a preprocessing step for
filtering the barcodes and normalizing the count values.

Internal function used in \code{SingleCellAlleleExperiment()} constructor as a preprocessing step for
filtering the barcodes and normalizing the count values.
}
