% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{find_allele_ids}
\alias{find_allele_ids}
\title{\code{find_allele_ids} internal function used to subsample the quantification assay and only
return the rows specifying allele-quantification information}
\usage{
find_allele_ids(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{
subsample of the sce containing all rows with allele_counts
}
\description{
\code{find_allele_ids} internal function used to subsample the quantification assay and only
return the rows specifying allele-quantification information
}
