% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{find_allele_ids}
\alias{find_allele_ids}
\title{Identify rows containing allele information}
\usage{
find_allele_ids(sce, exp_type)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{exp_type}{character string to pass information about the used experimental approach. Either use \code{exp_type = "WTA"} or \code{exp_type = "Amplicon"}. Passed from \code{SingleCellAlleleExperiment()}.}
}
\value{
subsample of the scae containing all rows with allele_counts
}
\description{
Internal function used in \code{get_allelecounts()} to subsample the quantification assay and only
return the rows specifying allele-quantification information.
}
