% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAlleleExperiment.R
\name{find_allele_ids}
\alias{find_allele_ids}
\title{Identify rows containing allele information for WTA}
\usage{
find_allele_ids(sce)
}
\arguments{
\item{sce}{A \code{\link{SingleCellExperiment}} object.}
}
\value{
A SingleCellExperiment object.
}
\description{
Internal function used in \code{get_allelecounts()} to subsample the quantification assay and only
return the rows specifying allele-quantification information.
}
