% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{find_not_ident}
\alias{find_not_ident}
\title{Find not yet known allele identifiers}
\usage{
find_not_ident(scae, agene_names)
}
\arguments{
\item{scae}{SingleCellAlleleExperiment object}

\item{agene_names}{list of character strings with allele gene names}
}
\value{
list of identifiers that can not be found in the allele lookup table
}
\description{
Internal function used in \code{get_allelecounts()}to find allele identifier that arent present in the lookup table.
}
