% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{genes2functional}
\alias{genes2functional}
\title{Building second new subassay for the SingleCellAlleleExperiment object}
\usage{
genes2functional(sce, lookup, exp_type)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{lookup}{lookup table info passed from \code{SingleCellAlleleExperiment()}.}

\item{exp_type}{character string to pass information about the used experimental approach. Either use \code{exp_type = "WTA"} or \code{exp_type = "Amplicon"}, passed from the \code{readAlleleCounts()} function.}
}
\value{
SingleCellAlleleExperiment object
}
\description{
Internal function for the second assay extension used in the \code{SingleCellAlleleExperiment()} constructor
computing the second of the two new subassays that get appended to the
quantification assay. This subassay contains the functional allele classes and
sums up the expression counts of the allele genes that are in the same functional group.
}
