% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{genes2functional}
\alias{genes2functional}
\title{Building second new subassay for the SingleCellAlleleExperiment object}
\usage{
genes2functional(sce, lookup, exp_type)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{lookup}{allele lookup table}

\item{exp_type}{either "WTA" or "Amplicon" depending on the used experiments technology}
}
\value{
adds functional subassay containing summarized count information for the functional allele classes
}
\description{
Internal function for the second assay extension used in the \code{SingleCellAlleleExperiment()} constructor
computing the second of the two new subassays that get appended to the
quantification assay. This subassay contains the functional allele classes and
sums up the expression counts of the allele genes that are in the same functional group.
}
