% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAlleleExperiment.R
\name{genes2functional}
\alias{genes2functional}
\title{Building second new subassay for the SingleCellAlleleExperiment object}
\usage{
genes2functional(sce, lookup, exp_type)
}
\arguments{
\item{sce}{A \code{\link{SingleCellExperiment}} object.}

\item{lookup}{A data.frame object containing the lookup table.}

\item{exp_type}{A character string determining whether the gene symbols in the input data are Ensemble identifiers or ncbi identifiers. Only used internally, not related to input done by the user.}
}
\value{
A SingleCellExperiment object.
}
\description{
Internal function for the second assay extension used in the \code{SingleCellAlleleExperiment()} constructor
computing the second of the two new subassays that get appended to the
quantification assay. This subassay contains the functional allele classes and
sums up the expression counts of the allele genes that are in the same functional group.
}
