% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{genes2functional}
\alias{genes2functional}
\title{\code{genes2functional} second internal translation function used in the SingleCellAlleleExperiment constructor
computing the second of the two new subassays that get appended to the
quantification assay. This subassay contains the functional allele classes and
sums up the expression counts of the allele genes that are in the same functional group}
\usage{
genes2functional(sce, lookup)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{lookup}{allele lookup table}
}
\value{
adds functional subassay containing summarized count information for the functional allele classes
}
\description{
\code{genes2functional} second internal translation function used in the SingleCellAlleleExperiment constructor
computing the second of the two new subassays that get appended to the
quantification assay. This subassay contains the functional allele classes and
sums up the expression counts of the allele genes that are in the same functional group
}
