% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_agenes}
\alias{get_agenes}
\title{Get allele gene rows}
\usage{
get_agenes(scae)
}
\arguments{
\item{scae}{SingleCellAlleleExperiment object}
}
\value{
subsampled SingleCellAlleleExperiment object
}
\description{
Getter function returning subsampled SCAE-object with all rows containing allele gene information. These rows are
identfied by "I" in rowData(scae)$NI_I and "G" in rowData(scae)$Quant_type.
}
