% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_agenes}
\alias{get_agenes}
\title{Get immune gene rows}
\usage{
get_agenes(scae)
}
\arguments{
\item{scae}{SingleCellAlleleExperiment object}
}
\value{
subsampled SingleCellAlleleExperiment object
}
\description{
Getter function returning subsampled \verb{SCAE object} with all rows containing immune gene information. These rows are
identfied by \code{I} in \code{rowData(scae)$NI_I} and \code{G} in \code{rowData(scae)$Quant_type} columns.
}
