% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_agenes}
\alias{get_agenes}
\title{Get immune gene rows}
\usage{
get_agenes(scae)
}
\arguments{
\item{scae}{A \code{\link{SingleCellAlleleExperiment}} object.}
}
\value{
A SingleCellAlleleExperiment object.
}
\description{
Getter function returning subsampled SCAE object with all rows containing immune gene information. These rows are
identfied by "I" in rowData(scae)$NI_I and "G" in rowData(scae)$Quant_type.
}
