% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{get_allelecounts}
\alias{get_allelecounts}
\title{Get Subassay with allele gene names and raw allele quantification}
\usage{
get_allelecounts(sce, lookup, exp_type)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{lookup}{allele lookup table}

\item{exp_type}{either "WTA" or "Amplicon" depending on the used experiments technology}
}
\value{
subsample of the sce containing all allele counts with the allele gene identifier as rownames
}
\description{
Internal function used to build a subassay containing counts from raw alleles
The rownames  of this subassay are already translated to the corresponding allele gene identifier, which
are extracted from the allele lookup table
}
