% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAlleleExperiment.R
\name{get_allelecounts}
\alias{get_allelecounts}
\title{Get Subassay with allele gene names and raw allele quantification}
\usage{
get_allelecounts(sce, lookup, exp_type)
}
\arguments{
\item{sce}{A \code{\link{SingleCellExperiment}} object.}

\item{lookup}{A data.frame object containing the lookup table.}

\item{exp_type}{A vector containing two character strings. Either \code{"WTA"} or \code{"Amplicon"} are valid inputs. Choose one depending on the used transcriptomics approach.}
}
\value{
A SingleCellExperiment object.
}
\description{
Internal function used to build a subassay containing counts from raw alleles
The rownames  of this subassay are already translated to the corresponding allele gene identifier, which
are extracted from the allele lookup table
}
