% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{get_allelecounts}
\alias{get_allelecounts}
\title{\code{get_allelecounts} internal function used to build a subassay containing only the allelecounts.
The rownames  of this subassay are already translated to the corresponding allele_gene names, which
are extracted from the allele lookup table}
\usage{
get_allelecounts(sce, lookup)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{lookup}{allele lookup table}
}
\value{
subsample of the sce containing all allele_counts with the allele_gen_names as rownames
}
\description{
\code{get_allelecounts} internal function used to build a subassay containing only the allelecounts.
The rownames  of this subassay are already translated to the corresponding allele_gene names, which
are extracted from the allele lookup table
}
