% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_alleles}
\alias{get_alleles}
\title{Get allele rows}
\usage{
get_alleles(scae)
}
\arguments{
\item{scae}{SingleCellAlleleExperiment object}
}
\value{
subsampled SingleCellAlleleExperiment object
}
\description{
Getter function returning subsampled \verb{SCAE object} with all rows containing raw allele information. These rows are
identified by \code{I} in \code{rowData(scae)$NI_I} and \code{A} in \code{rowData(scae)$Quant_type} columns.
}
