% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_alleles}
\alias{get_alleles}
\title{Get allele rows}
\usage{
get_alleles(scae)
}
\arguments{
\item{scae}{SingleCellAlleleExperiment object}
}
\value{
subsampled SingleCellAlleleExperiment object
}
\description{
Getter function returning subsampled SCAE object with all rows containing raw allele information. These rows are
identified by "I" in rowData(scae)$NI_I and "A" in rowData(scae)$Quant_type.
}
