% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{get_func}
\alias{get_func}
\title{\code{get_func} getter function which creates scae-object for the subassay containing functional class information}
\usage{
get_func(scae)
}
\arguments{
\item{scae}{SingleCellAlleleExperiment object}
}
\value{
subsampled SingleCellAlleleExperiment object
}
\description{
\code{get_func} getter function which creates scae-object for the subassay containing functional class information
}
