% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_func}
\alias{get_func}
\title{Get functional class rows}
\usage{
get_func(scae)
}
\arguments{
\item{scae}{SingleCellAlleleExperiment object}
}
\value{
subsampled SingleCellAlleleExperiment object
}
\description{
Getter function returning subsampled \verb{SCAE object} with all rows containing functional class information. These rows are
identified by \code{I} in \code{rowData(scae)$NI_I} and \code{F} in \code{rowData(scae)$Quant_type} columns.
}
