% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKnee.R
\name{get_knee_info}
\alias{get_knee_info}
\title{Knee plot}
\usage{
get_knee_info(matrix, genes, barcodes)
}
\arguments{
\item{matrix}{A sparse \code{\link{Matrix}} object containing the quantification data.}

\item{genes}{A data.frame object containing gene identifiers.}

\item{barcodes}{A data.frame object containing barcode identifiers.}
}
\value{
A list including a data.frame with barcode rank information, the corresponding knee and inflection point.
}
\description{
Creates a knee plot ranking the barcodes according to their total UMI count. The plot is
used to determine a threshold for filtering barcodes in the preprocessing step.
}
