% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{get_ncbi_gene_names}
\alias{get_ncbi_gene_names}
\title{Get Ncbi genes using biomaRt}
\usage{
get_ncbi_gene_names(ensembl_ids)
}
\arguments{
\item{ensembl_ids}{vector containing ensembl.ids}
}
\value{
vector containing the ncbi names matching the right rows for the ensembl.ids
}
\description{
Get Ncbi genes using biomaRt
}
