% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAlleleExperiment.R
\name{get_ncbi_gene_names}
\alias{get_ncbi_gene_names}
\title{Get NCBI genes using biomaRt}
\usage{
get_ncbi_gene_names(ensembl_ids)
}
\arguments{
\item{ensembl_ids}{A vector containing character strings for ensemble gene identifiers.}
}
\value{
A vector containing character strings for NCBI gene names.
}
\description{
This internal function is used to retrieve the gene-symbol names to the corresponding ENSG accession numbers in the WTA experiment approach.
Internet connection is mandatory, as its retrieving the newest possible dataset every time. If you have to work offline then use the
get_ncbi_org by specifying \code{symbols = "orgdb"} in the corresponding \code{symbols} parameter of the \code{readAlleleCounts()} function.
}
