% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAlleleExperiment.R
\name{get_ncbi_org}
\alias{get_ncbi_org}
\title{Get NCBI genes using the org.HS.db package}
\usage{
get_ncbi_org(sce)
}
\arguments{
\item{sce}{A \code{\link{SingleCellExperiment}} object.}
}
\value{
A list of character strings for gene names.
}
\description{
Get NCBI genes using the org.HS.db package
}
