% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_nigenes}
\alias{get_nigenes}
\title{Get non-immune gene rows}
\usage{
get_nigenes(scae)
}
\arguments{
\item{scae}{SingleCellAlleleExperiment object}
}
\value{
subsampled SingleCellAlleleExperiment object
}
\description{
Getter function returning subsampled \verb{SCAE object} with all rows containing non immune gene information. These rows are
identified by \code{NI} in \code{rowData(scae)$NI_I} and \code{G} in \code{rowData(scae)$Quant_type} columns.
}
