% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_nigenes}
\alias{get_nigenes}
\title{Get non-immune rows}
\usage{
get_nigenes(scae)
}
\arguments{
\item{scae}{SingleCellAlleleExperiment object}
}
\value{
subsampled SingleCellAlleleExperiment object
}
\description{
Getter function returning subsampled SCAE object with all rows containing non immune gene information. These rows are
identified by "NI" in rowData(scae)$NI_I and "G" in rowData(scae)$Quant_type.
}
