% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAlleleExperiment.R
\name{log_transform}
\alias{log_transform}
\title{Log-transform normalized counts}
\usage{
log_transform(sce)
}
\arguments{
\item{sce}{A \code{\link{SingleCellExperiment}} object.}
}
\value{
A SingleCellExperiment object.
}
\description{
Internal function used in the \code{SingleCellAlleleExperiment()} constructor to log-normalize the raw counts and add them to the \code{\link{logcounts}} assay.
}
