% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{log_transform}
\alias{log_transform}
\title{Log-transform normalized counts}
\usage{
log_transform(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{
SingleCellAlleleExperiment object with an additional assay containing
}
\description{
Internal function used in the \code{SingleCellAlleleExperiment()} constructor to log-normalize the raw counts and add them to \code{logcounts} assay.
}
