% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKnee.R
\name{plotKnee}
\alias{plotKnee}
\title{Knee plot}
\usage{
plotKnee(matrix_file, gene_file, barcode_file)
}
\arguments{
\item{matrix_file}{count matrix that was read in in the \code{read_from_sparse_allele()} function}

\item{genes_file}{list with gene identifiers that was read in in the \code{read_from_sparse_allele()} function}

\item{barcodes_file}{list with barcode identifiers that was read in in the \code{read_from_sparse_allele()} function}
}
\value{
returns a knee plot for determining a count threshold used for filtering out barcodes
}
\description{
Creates a knee plot ranking the barcodes according to their total UMI count. The plot is
used to determine a threshold for filtering barcodes in the preprocessing step.
}
