% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKnee.R
\name{plotKnee}
\alias{plotKnee}
\title{Knee plot}
\usage{
plotKnee(matrix_file, genes_file, barcodes_file)
}
\arguments{
\item{matrix_file}{file path of the directory containing the input files as character string}

\item{genes_file}{file path of the directory containing the input files as character string}

\item{barcodes_file}{file path of the directory containing the input files as character string}
}
\value{
returns a knee plot for determining a count threshold used for filtering out barcodes
}
\description{
Creates a knee plot ranking the barcodes according to their total UMI count. The plot is
used to determine a threshold for filtering barcodes in the preprocessing step.
}
