% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlleleCounts.R
\name{read_from_sparse_allele}
\alias{read_from_sparse_allele}
\title{Reading in allele-aware quantification data}
\usage{
read_from_sparse_allele(path, exp_type = exp_type)
}
\arguments{
\item{path}{character string of the file path to the directory containing the expected input files}

\item{exp_type}{character string to pass information about the used experimental approach. Either use \code{exp_type = "WTA"} or \code{exp_type = "Amplicon"}.}
}
\value{
list with the read_in data sorted into different slots
}
\description{
Internal function used in \code{readAlleleCounts()} that reads in the data stated in the given directory path.
}
