% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlleleCounts.R
\name{read_from_sparse_allele}
\alias{read_from_sparse_allele}
\title{Reading in allele-aware quantification data}
\usage{
read_from_sparse_allele(
  path,
  exp_type = exp_type,
  barcode_file,
  gene_file,
  matrix_file
)
}
\arguments{
\item{path}{A character string determining the path to the directory containing the input files.}

\item{exp_type}{A vector containing two character strings. Either \code{"WTA"} or \code{"Amplicon"} are valid inputs. Choose one depending on the used transcriptomics approach.}

\item{barcode_file}{A character string determining the name of the file containing the sample-tag quantification data.}

\item{gene_file}{A character string determining the name of the file containing the feature identifiers.}

\item{matrix_file}{A character string determining the name of the file containing the count matrix.}
}
\value{
A list with three data.frames containing the input data information.
}
\description{
Internal function used in \code{read_allele_counts()} that reads in the data stated in the given directory path.
}
