% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlleleCounts.R
\name{read_Lookup}
\alias{read_Lookup}
\title{Read in allele lookup}
\usage{
read_Lookup(path, lookup_file)
}
\arguments{
\item{path}{A character string determining the path to the directory containing the input files.}

\item{lookup_file}{A character string determining the name of the lookup table file.}
}
\value{
A data.frame containing a representation of the lookup table.
}
\description{
Internal function used in \code{read_allele_counts()} to read in the allele lookup table.
}
