% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlleleCounts.R
\name{readAlleleCounts}
\alias{readAlleleCounts}
\title{\code{readAlleleCounts} main read_in function for reading in the allele quantification data and
loading the data into an SingleCellAlleleExperiment object. Input data are stored in a shared folder.
Expected naming scheme of the files: quantification matrix: "matrix.mtx"
barcode information: "barcodes.txt"
feature information: "features.txt"
allele lookup table: "lookup_table_HLA_only"}
\usage{
readAlleleCounts(
  samples,
  sample.names = names(samples),
  col.names = TRUE,
  type = c("auto", "sparse", "HDF5", "prefix"),
  compressed = NULL,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{samples}{character string input containing the path to the directory containing the
input files}

\item{sample.names}{character string for a sample_name identifier}

\item{col.names}{binary variable indicating whether quantification assay should contain column names}

\item{type}{vector of character strings containing options for the input data_type}

\item{compressed}{binary variable whether the input files are .gz compressed}

\item{BPPARAM}{A BiocParallelParam object specifying how loading should be parallelized for multiple samples}
}
\value{
SingleCellAlleleExperiment object
}
\description{
\code{readAlleleCounts} main read_in function for reading in the allele quantification data and
loading the data into an SingleCellAlleleExperiment object. Input data are stored in a shared folder.
Expected naming scheme of the files: quantification matrix: "matrix.mtx"
barcode information: "barcodes.txt"
feature information: "features.txt"
allele lookup table: "lookup_table_HLA_only"
}
