% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlleleCounts.R
\name{readAlleleCounts}
\alias{readAlleleCounts}
\title{Reading in allele quantification data into SingleCellAlleleExperiment object}
\usage{
readAlleleCounts(
  samples,
  sample.names = names(samples),
  col.names = TRUE,
  compressed = NULL,
  filter_threshold = 0,
  exp_type = c("WTA", "Amplicon"),
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{samples}{character string input containing the path to the directory containing the
input files}

\item{sample.names}{character string for a sample_name identifier}

\item{col.names}{binary variable indicating whether quantification assay should contain column names}

\item{compressed}{binary variable whether the input files are .gz compressed}

\item{filter_threshold}{count threshold for filtering barcodes/cells}

\item{exp_type}{either "WTA" or "Amplicon" depending on the used experiments technology}

\item{BPPARAM}{A BiocParallelParam object specifying how loading should be parallelized for multiple samples}
}
\value{
SingleCellAlleleExperiment object
}
\description{
Main read in function for reading in given allele quantification data and
loading the data into an SingleCellAlleleExperiment object. Input data are stored in a shared folder.
Expected naming scheme of the files: quantification matrix: "matrix.mtx"
barcode information: "barcodes.txt"
feature information: "features.txt"
allele lookup table: "lookup_table_HLA_only"
}
