% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlleleCounts.R
\name{readAlleleCounts}
\alias{readAlleleCounts}
\title{Reading in allele quantification data into SingleCellAlleleExperiment object}
\usage{
readAlleleCounts(
  samples,
  sample_names = names(samples),
  filter = c("yes", "no", "custom"),
  exp_type = c("WTA", "Amplicon"),
  symbols = NULL,
  lookup_file = "lookup_table_HLA_only.txt",
  barcode_file = "cells_x_genes.barcodes.txt",
  gene_file = "cells_x_genes.genes.txt",
  matrix_file = "cells_x_genes.mtx",
  tag_feature_mtx = "cells_x_features.mtx",
  tag_feature_barcodes = "cells_x_features.barcodes.txt",
  filter_threshold = NULL,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{samples}{character string input containing the path to the directory containing the
input files}

\item{sample_names}{character string for a sample_name identifier}

\item{filter}{character string determining the filter mode. \code{yes} uses the inflection point of the knee plot. \code{no} computes the knee plot and stops funciton execution. \code{custom} allows for setting a custom threshold in \code{filter_threshold}.}

\item{exp_type}{either \code{WTA} or \code{Amplicon} depending on the used experiments technology}

\item{symbols}{identifier used to choose which database-function to use for retrieving the ncbi gene names}

\item{lookup_file}{character string determining the name of the lookup table file}

\item{barcode_file}{character string determining the name of the file containing the barcode identifiers}

\item{gene_file}{character string determining the name of the file containing the feature identifiers}

\item{matrix_file}{character string determining the name of the file containing the count matrix}

\item{tag_feature_mtx}{character string determining the name of the file containing the count matrix of sample tag information}

\item{tag_feature_barcodes}{character string determining the name of the file containing the barcode identifiers of sample tag information}

\item{filter_threshold}{NULL or integer value > 0 used for \code{custom} filtering if \code{filter = "custom"}}

\item{BPPARAM}{A BiocParallelParam object specifying how loading should be parallelized for multiple samples}
}
\value{
SingleCellAlleleExperiment object
}
\description{
Main read in function for reading in given allele quantification data and
loading the data into an \code{SingleCellAlleleExperiment} object. Input data are stored in a shared folder.
Expected naming scheme of the files:
\itemize{
\item quantification matrix: \code{matrix.mtx}
\item barcode information: \code{barcodes.txt}
\item feature information: \code{features.txt}
\item allele lookup table: \code{lookup_table_HLA_only}
}
}
