% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlleleCounts.R
\name{readLookup}
\alias{readLookup}
\title{\code{readLookup} internal function to read in the lookup table}
\usage{
readLookup(path)
}
\arguments{
\item{path}{character string input containing the path to the directory containing the
input files; helper function used in readAlleleCounts()}
}
\value{
lookup table
}
\description{
\code{readLookup} internal function to read in the lookup table
}
