% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlleleCounts.R
\name{readLookup}
\alias{readLookup}
\title{Read in allele lookup}
\usage{
readLookup(path)
}
\arguments{
\item{path}{file path of the directory containing the input files as character string}
}
\value{
lookup table
}
\description{
Internal function used in \code{readAlleleCounts()} to read in the allele lookup table.
}
