% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlleleCounts.R
\name{readLookup}
\alias{readLookup}
\title{Read in allele lookup}
\usage{
readLookup(path, exp_type, lookup_file)
}
\arguments{
\item{path}{file path of the directory containing the input files as character string}

\item{exp_type}{either "WTA" or "Amplicon" depending on the used experiments technology}

\item{lookup_file}{character string determining the name of the lookup table file}
}
\value{
lookup table
}
\description{
Internal function used in \code{readAlleleCounts()} to read in the allele lookup table.
}
