% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAlleleExperiment-rowData.R
\name{rowData<-,SingleCellAlleleExperiment-method}
\alias{rowData<-,SingleCellAlleleExperiment-method}
\title{Setting rowData}
\usage{
\S4method{rowData}{SingleCellAlleleExperiment}(x, ...) <- value
}
\arguments{
\item{x}{A \code{\link{SingleCellAlleleExperiment}} object}

\item{...}{additional parameters passed to rowData<-}

\item{value}{Value of valid type and content (see validty.R)}
}
\value{
A \code{\link{SingleCellAlleleExperiment}} object
}
\description{
Function used to change and set new values in the \code{rowData} slot.
Incorporates the validity checks on the output SCAE object to catch
invalid changes.
}
