% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{scae_subset}
\alias{scae_subset}
\title{Subset SCAE object}
\usage{
scae_subset(
  scae,
  subset = c("nonimmune", "alleles", "immune_genes", "functional", "unknown_alleles")
)
}
\arguments{
\item{scae}{SCAE object}

\item{subset}{character string specifiying a data layer}
}
\value{
SCAE object
}
\description{
Subset SCAE object
}
