% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{scae_subset_functional}
\alias{scae_subset_functional}
\title{Get functional class rows}
\usage{
scae_subset_functional(scae)
}
\arguments{
\item{scae}{A \code{\link{SingleCellAlleleExperiment}} object.}
}
\value{
A SingleCellAlleleExperiment object.
}
\description{
Getter function returning subsampled SCAE object with all rows containing functional class information. These rows are
identified by "I" in rowData(scae)$NI_I and "F" in rowData(scae)$Quant_type.
}
