% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{scae_subset_functional}
\alias{scae_subset_functional}
\title{Get functional class rows}
\usage{
scae_subset_functional(scae)
}
\arguments{
\item{scae}{A \code{\link{SingleCellAlleleExperiment}} object.}
}
\value{
A SingleCellAlleleExperiment object.
}
\description{
Getter function returning subsampled SCAE object with all rows containing functional class information. These rows are
identified by "I" in rowData(scae)$NI_I and "F" in rowData(scae)$Quant_type.
}
\examples{

example_data <- system.file("extdata", package = "SingleCellAlleleExperiment")

scae <- read_allele_counts(example_data,
                        sample_names = "example_data",
                        filter = "yes",
                        exp_type = "WTA",
                        lookup_file = "lookup_table_HLA_only.csv",
                        barcode_file = "cells_x_genes.barcodes.txt",
                        gene_file = "cells_x_genes.genes.txt",
                        matrix_file = "cells_x_genes.mtx",
                        tag_feature_mtx = "cells_x_genes.genes.txt",
                        tag_feature_barcodes = "cells_x_genes.barcodes.txt",
                        filter_threshold = NULL)

scae

rownames(scae_subset_functional(scae))

scae_functional_class <- scae_subset_functional(scae)

scae_functional_class

rownames(scae_functional_class)

}
