% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAlleleExperiment.R
\docType{class}
\name{SingleCellAlleleExperiment-class}
\alias{SingleCellAlleleExperiment-class}
\alias{.scae}
\title{SingleCellAlleleExperiment class definition}
\value{
definition for the scae class
}
\description{
Defining the \code{SingleCellAlleleExperiment} class derived from \code{SingleCellExperiment} class.
}
