% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{SingleCellAlleleExperiment}
\alias{SingleCellAlleleExperiment}
\title{Constructor SingleCellAlleleExperiment class}
\usage{
SingleCellAlleleExperiment(..., threshold, exp_type, symbols, lookup)
}
\arguments{
\item{...}{parameters to pass to SingleCellExperiment constructor}

\item{threshold}{count threshold for filtering barcodes/cells}

\item{exp_type}{either \code{"WTA"} or \code{"Amplicon"} depending on the used experiments technology}

\item{symbols}{identifier used to choose which database-function to use to retrieve the ncbi gene names}

\item{lookup}{allele lookup file}
}
\value{
SingleCellAlleleExperiment object
}
\description{
Constructor for the \code{SingleCellAllelExperiment} (SCAE) class.
Constructor is used in the read in function \code{readAlleleCounts()}. Performing all necessary steps to transform
a \code{SingleCellExperiment} object into the extended \code{SingleCellAlleleExperiment} object. SCAE objects
contain an extended count-assay aswell as extended rowData.
}
