% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{SingleCellAlleleExperiment}
\alias{SingleCellAlleleExperiment}
\title{Constructor SingleCellAlleleExperiment class}
\usage{
SingleCellAlleleExperiment(
  ...,
  threshold,
  exp_type,
  symbols,
  lookup,
  verbose = FALSE
)
}
\arguments{
\item{...}{parameters to pass to SingleCellExperiment constructor}

\item{threshold}{count threshold for filtering barcodes/cells}

\item{exp_type}{either \code{"WTA"} or \code{"Amplicon"} depending on the used experiments technology}

\item{symbols}{identifier used to choose which database-function to use to retrieve the ncbi gene names}

\item{lookup}{allele lookup file}

\item{verbose}{FALSE if no info message for runtime should be shown (default), TRUE for runtime information about each step}
}
\value{
SingleCellAlleleExperiment object
}
\description{
Constructor for the \code{SingleCellAllelExperiment} (SCAE) class.
Constructor is used in the read in function \code{readAlleleCounts()}. Performing all necessary steps to transform
a \code{SingleCellExperiment} object into the extended \code{SingleCellAlleleExperiment} object. SCAE objects
contain allele, gene and functional level quantification results. The additional layers are stored as additional
rows in the count assays as well as in extended rowData.
}
