% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellAlleleExperiment.R
\name{SingleCellAlleleExperiment}
\alias{SingleCellAlleleExperiment}
\title{Constructor SingleCellAlleleExperiment class}
\usage{
SingleCellAlleleExperiment(..., threshold, exp_type, lookup, verbose = FALSE)
}
\arguments{
\item{...}{Arguments passed to the \code{\link{SingleCellExperiment}} constructor to fill the slots of the SCE-class.}

\item{threshold}{An integer value used as a threshold for filtering low-quality barcodes/cells.}

\item{exp_type}{A vector containing two character strings. Either \code{"WTA"} or \code{"Amplicon"} are valid inputs. Choose one depending on the used transcriptomics approach.}

\item{lookup}{A data.frame object containing the lookup table.}

\item{verbose}{A logical parameter to decide if runtime-messages should be shown during function execution.
Use \code{FALSE} if no info runtime-messages should be shown (default), and \code{TRUE} for showing runtime-messages.}
}
\value{
A SingleCellAlleleExperiment object.
}
\description{
Constructor for the \code{SingleCellAllelExperiment} (SCAE) class.
Constructor is used in the read in function \code{read_allele_counts()}. Performing all necessary steps to transform
a \code{SingleCellExperiment} object into the extended \code{SingleCellAlleleExperiment} object. SCAE objects
contain allele, gene and functional level quantification results. The additional layers are stored as additional
rows in the count assays as well as in extended rowData.
}
