% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecellalleleexperiment.R
\name{SingleCellAlleleExperiment}
\alias{SingleCellAlleleExperiment}
\title{\code{SingleCellAlleleExperiment} Constructor for the scae class}
\usage{
SingleCellAlleleExperiment(..., lookup)
}
\arguments{
\item{...}{parameters to fill in the sce constructor}

\item{lookup}{allele lookup file}
}
\value{
SingleCellAlleleExperiment object
}
\description{
\code{SingleCellAlleleExperiment} Constructor for the scae class
}
