% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/SingleCellAlleleExperiment.R
\docType{class}
\name{SingleCellAlleleExperiment-class}
\alias{SingleCellAlleleExperiment-class}
\alias{SingleCellAlleleExperiment}
\title{Constructor SingleCellAlleleExperiment class}
\usage{
SingleCellAlleleExperiment(
  ...,
  lookup,
  metadata = NULL,
  threshold = 0,
  exp_type = "ENS",
  log = FALSE,
  gene_symbols = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{...}{Arguments passed to the \code{\link{SingleCellExperiment}} constructor to fill the slots of the SCE-class.}

\item{lookup}{A data.frame object containing the lookup table.}

\item{metadata}{potential information regarding plotting a knee plot for quality control.}

\item{threshold}{An integer value used as a threshold for filtering low-quality barcodes/cells.}

\item{exp_type}{A vector containing two character strings. Either \code{"WTA"} or \code{"Amplicon"} are valid inputs. Choose one depending on the used transcriptomics approach.}

\item{log}{binary if user wants to compute \code{logcounts} assay.}

\item{gene_symbols}{A logical parameter to decide whether to compute the NCBI gene names in case the raw data only contains ENSEMBLE gene identifiers.}

\item{verbose}{A logical parameter to decide if runtime-messages should be shown during function execution.
Use \code{FALSE} if no info runtime-messages should be shown (default), and \code{TRUE} for showing runtime-messages.}
}
\value{
A SingleCellAlleleExperiment object.
}
\description{
Constructor for the \code{SingleCellAllelExperiment} (SCAE) class.
Constructor is used in the read in function \code{read_allele_counts()}. Performing all necessary steps to transform
a \code{SingleCellExperiment} object into the extended \code{SingleCellAlleleExperiment} object. SCAE objects
contain allele, gene and functional level quantification results. The additional layers are stored as additional
rows in the count assays as well as in extended rowData.
}
